config/urls.py
from django.contrib import admin
from django.urls import include, path
from debug_toolbar.toolbar import debug_toolbar_urls

from config.django import base

urlpatterns = [
    path("admin/", admin.site.urls),
    path("api-auth/", include("rest_framework.urls")),
]

if base.DEBUG:
    urlpatterns.extend(debug_toolbar_urls())